FPLA-2043: e2e tests updated
