+++ e2e/tests/issueResolution_test.js
@@ -1,5 +1,6 @@
 const config = require('../config.js');
 const issueResolution = require('../fixtures/IssueResolution.json');
+const cmoHelper = require('../helpers/cmo_helper');
 
 let caseId;
 
@@ -9,29 +10,22 @@ BeforeSuite(async (I) => {
   caseId = await I.submitNewCaseWithData(issueResolution);
 });
 
-Scenario('Issue resolution', async (I, caseViewPage, uploadCaseManagementOrderEventPage, reviewAgreedCaseManagementOrderEventPage) => {
+Scenario('Judge transitions CMO to issue resolution case date', async (I, caseViewPage, caseListPage, uploadCaseManagementOrderEventPage, reviewAgreedCaseManagementOrderEventPage) => {
   await I.navigateToCaseDetailsAs(config.swanseaLocalAuthorityUserOne, caseId);
-  await localAuthoritySendsAgreedCmo(I, caseViewPage, uploadCaseManagementOrderEventPage, '1 January 2020', true);
+  await cmoHelper.localAuthoritySendsAgreedCmo(I, caseViewPage, uploadCaseManagementOrderEventPage, '1 January 2020', true);
   I.seeEventSubmissionConfirmation(config.applicationActions.uploadCMO);
-  await localAuthoritySendsAgreedCmo(I, caseViewPage, uploadCaseManagementOrderEventPage);
+  await cmoHelper.localAuthoritySendsAgreedCmo(I, caseViewPage, uploadCaseManagementOrderEventPage);
   I.seeEventSubmissionConfirmation(config.applicationActions.uploadCMO);
   await I.navigateToCaseDetailsAs(config.judicaryUser, caseId);
   await caseViewPage.goToNewActions(config.applicationActions.reviewAgreedCmo);
+  reviewAgreedCaseManagementOrderEventPage.selectCMOToReview('1 January 2020');
+  await I.retryUntilExists(() => I.click('Continue'), '#reviewCMODecision_decision');
   reviewAgreedCaseManagementOrderEventPage.selectSealCmo();
   await I.completeEvent('Save and continue', {summary: 'Summary', description: 'Description'});
   I.seeEventSubmissionConfirmation(config.applicationActions.reviewAgreedCmo);
+  caseListPage.navigate();
+  caseListPage.changeStateFilter('Issue Resolution');
+  I.click(caseListPage.locateCase(caseId));
 });
-
-const localAuthoritySendsAgreedCmo = async function (I, caseViewPage, uploadCaseManagementOrderEventPage, hearingDate, multiHearings) {
-  await caseViewPage.goToNewActions(config.applicationActions.uploadCMO);
-
-  if (multiHearings) {
-    await uploadCaseManagementOrderEventPage.associateHearing(hearingDate);
-    await I.retryUntilExists(() => I.click('Continue'), '#uploadedCaseManagementOrder');
-  }
-
-  await uploadCaseManagementOrderEventPage.uploadCaseManagementOrder(config.testNonEmptyWordFile);
-  await I.completeEvent('Submit');
-};
